Update docs with optional empty quote_char
When managing CSV style files without quoting (which although perhaps invalid, still happens) it is important to note that with the **force_simple_split: true** option, you CAN set the **quote_char** to be empty. If you try to set the quote_char to empty without the simple_split, it will throw an exception.